requirements.txt
unstructured
markdown
SpeechRecognition
pyaudio
python-dotenv
pyttsx3
langchain
openai
